LayoutTests/imported/w3c/web-platform-tests/html/browsers/browsing-the-web/back-forward-cache/eligibility/broadcast-channel-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: TypeError: Invalid URL

Harness Error (TIMEOUT), message = null

TIMEOUT Eligibility (BroadcastChannel) Test timed out
NOTRUN Eligibility (BroadcastChannel closed in the pagehide event)
